Refactor into vmrunner
